+++ src/main/java/com/javarush/island/ogarkov/repository/itemfactory/Factory.java
@@ -8,7 +8,7 @@ public interface Factory {
 
     void addCreatedItem(Items item);
 
-    long getCreatedItemsCount();
+    int getCreatedItemsCount();
 
 
 }
